Entregas - Clara Polizel/ADK Agents/requirements.txt
google-adk
python-dotenv
google-genai